feat(nvim): Update Copilot and plugin configurations
This commit introduces several changes to the Neovim configuration. It
adds new configurations for the CopilotChat and Copilot plugins, and
removes the old Copilot configuration. It also updates the lazy-lock.json
file with new commit hashes for the gitsigns.nvim and nvim-treesitter
plugins. Lastly, it modifies the mini.lua file to check if the GitHub
Copilot client is active and updates the statusline accordingly.